ZC Bcast API: Fix bug seen in smp mode with 2 level bcast (#3161)
The bug is seen as a crash on the child node when it tries to forward the
message to its peer nodes after receiving a bcast message forwarded by an
an intermediate root node. The bug was caused by an incorrect CMI_DEST_RANK
and the fix was to reset this on the child node before forwarding it to peers.