Add BeginReceiver and EndReceiver to CkSyncBarrier
This allows users to add callbacks to run before and after the normal
AtSync calbacks, which can be used for setup, cleanup, or to prepare
for launching callbacks or resuming normal execution. Namely, this is
used by CkLocMgr in order to tell the LB framework that it is starting
or stopping object insertion, which was previously done by shoehorning
it into being an AtSync client unnaturally.

Additionally, the flooding mechanism of AtSync, intended to cover
cases where there are empty PEs, can hang when the flood message is
received by a PE before that PE has exited its AtSync section, meaning
that its CkSyncBarrier is off. To fix this, track the epoch of the
flood message (now called "kick" to more closely match the semantics
of the procedure) so the barrier can check if the kick for the current
epoch has been received when the barrier is eventually enabled
again.