CkArray: Allow broadcasts of ckDestroy
Deliver array broadcasts to local elements in reverse order, starting at
the end of the local element vector and proceeding to the beginning.
This pattern allows the target function of the broadcast to destroy the
array element without invalidating the remainder of the iteration space
for the local element vector.